Adopt new Robolectric Activity style (per #5249)
