+++ AnkiDroid/src/test/AndroidManifest.xml
@@ -12,5 +12,11 @@
         <activity
             android:name="com.ichi2.anki.NullCollectionReviewer">
         </activity>
+        <activity
+            android:name="com.ichi2.anki.NonPagingCardTemplateEditor">
+        </activity>
+        <activity
+            android:name="com.ichi2.anki.TestPreviewer">
+        </activity>
     </application>
 </manifest>
