+++ AnkiDroid/src/test/java/com/ichi2/anki/TestPreviewer.java
@@ -1,0 +1,6 @@
+package com.ichi2.anki;
+
+public class TestPreviewer extends Previewer {
+    public boolean getShowingAnswer() { return mShowingAnswer; }
+    public void disableDoubleClickPrevention() { mLastClickTime = (AbstractFlashcardViewer.DOUBLE_TAP_IGNORE_THRESHOLD * -2); }
+}
